backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
llama-cpp-python
faster-whisper
soundfile
numpy
livekit-agents
livekit-plugins-silero
